Add django-reversion and test cdms integration
This adds django-reversion and a few basic tests to make sure that
saving versions and reverting them back do not cause any unwanted
cdms calls.

Versions are automatically saved transparently so that the dev doesn't
need to do anything.
When local objects are refreshed from cdms, the change is saved as an
extra revision.

Reverting deleted objects is not supported at the moment because of
some poor django-reversion design choices.
It's technically possibile to reimplement the method with custom logic
but not needed for now.