Improve tests depending on cdms modified value
The sync operations depend on the values of the modified fields being in
sync. Because of the use of .now() and because the tests run fast,
those values were usually the same but in some cases some tests would fail.

To resolve all problems, here we are changing the modified value returned by
the cdms api to be something completely different.
In this way we discovered a few of these problems which would have been
otherwise hidden.