fix passing envvars from bash to docker (+1 squashed commit) Squashed commits: [c63b529] remove unneeded config (+45 squashed commits) Squashed commits: [1c47564] fix envvars [58ccfe1] setting default value for database connection [bf6c417] Add delete_all function to client for cleaning up after tests
Also document the delete function.
[e71c39a] Assert that List can be generated for a single Account
[f7c7598] Assert service counter assertion helpers work with single instances
[083b0b2] Remove bootstrapped Account from DELETE test
[c1a7c86] Add basic test on CREATE with no data
[a4c30f8] Add two assertion helpers for checking numbers of instances
[ac61b59] Add bootstrapped tests on DELETE
[3b1c224] Increase verbosity on Circle test runs

Aim is to help with debugging of the skipIntegration decorator and to
give info during the test run on the reason that tests are skipped.
[37c331c] Test deletion of a missing instance

Also behaviour of 404s has been changed in the NTLM auth - they now
raise exceptions... This is one step closer to being the same
client-layer as AD auth.
[5cab0da] Extract Account instance tests to suite and test client to setUp
[a8d7ea9] Add info to README about settings for integration test
[c6549e1] Add first client layer integration test - list Organisations
[e70cd77] Add TEST_INTEGRATION Django setting, overriden by envvar
[faf072c] Add skipIntegration decorator for missing out integrated tests
[f22761e] Allow NTLM related settings to be overridden by envvars
[8da8d4e] Expand mocking shorthand to full names
[9a55156] Fix comment on Auth init - correct required settings
[3f47c30] Improve mock of ActiveDirectoryAuth in CDMSRestApi init

This means that just the constructor is checked that it was called and
an instance created of ActiveDirectoryAuth. This should be less brittle
than the previous strategy of checking that setup_session was called.
[4b0f1fa] Adjust NTLMAuth to return data from successful responses
[5f416a1] Allow for custom auth instance in CDMSRestApi
[2418246] Add make_request to NTLMAuth
[8c0f89b] Add stub NTLMAuth class with simple init
[3904730] Add requests_ntlm to base requirements
[2c00bf4] Fix mutable kwargs

First fix didn't work - `dict()` is still evaluated just once and
creates a mutable dict. Have to go for the full `None` and if statement
combination which I was trying to avoid.
[daa068e] Update README with further info about building requirements

Also:

* clean out the requirements files - I let Django extensions slip back
    in somehow.

* Add dependencies to the requirements Makefile so that Make understands
    which files depend on each other.
[4167cfe] Extract AD auth to a new auth module - create space for NTLM
[95ef2f1] Extract COOKIE_FILE to settings and test CookieStorageTestCase

`COOKIE_FILE` was module-level code which is hard to patch - it looks
more like a setting to me, so move it to settings.

Test that tearDown of CookieStorageTestCase is stable - it had been
failing often, especially after failed test runs.
[68171d1] Extract ActiveDirectory functions

Also clean up tests
[bacc605] Fix mutable kwargs
[429d82d] Extract ActiveDirectoryAuth. Rename BaseCDMSRestApiTestCase

BaseCDMSRestApiTestCase renamed to CookieStorageTestCase since its main
objective is to clean up cookie storage after tests.
[bb5e980] Simplify CDMSRestApi init tests (hopefully)
[fd74cbd] Move Active Directory tests to their own file. Fix flake8.

There will be a new set of tests for NTLM auth, so mark these existing
tests as focusing on AD auth.

Also `# flake8: noqa` meant that the whole of the file was being skipped
and not just the line. The line itself passes current flake8.
[f2a789b] Extract BaseCDMSRestApiTestCase
[e5e1e98] Adjust cookie cleaning to work at teardown time

This means that tests do not leave a cookie file on completion.
[22c505a] fix Dockerfile
[94e2a02] add db wait
[e608c81] simplify docker file
[c5cec93] fix pip install path
[95be0e9] should fix it?
[38b34e2] another try
[2074086] fix docker network on swarm cluster
[90f32af] fix docker network create in Jenkins job
[efb4f51] * fixes docker-compose network failure
* fix Jenkins job
* merge requirements (+1 squashed commit)
Squashed commits:
[5be99c5] port Marco's docker/jenkins definition